git-template: update git template regulation
With one day coding, I found the commit convention before is too ugly.
It is mixed with capitalized and non-capitalized message. The body is
sometime full of markup language sometime only contains plain text. Also
the input prompt is too hard to identify.

User should add the primary affected package as the title prefix, not
the file name. I found sometime you have to input capitalized file name
which will mess up one-line git log. Like modify for README.md, the
prefix may be "docs" or "readme".

Also from now on, all the title show not be capitalized. And I add a new
example to show how to write the title.

Besides, This patch moved guideline to the bottom of the template, and use
direction signs to show the input prompt.

For the body section, I think it is not useful to use list to say what
have been modified. It's better to use the commit form from
"Torvalds/Linux" and "golang/go". No markup, just write the body in
complete sentence like how we commenting in out code. With this
approach, the commit body can tell what's modified in a more rational
way.

Signed-off-by: Avimitin <avimitin@gmail.com>

Former-commit-id: 52daf9267d779ec0ab05ed7549eade0b8fb0aac1 [formerly a662bca93e5bffd1393624654ca3eda860d0878b]
Former-commit-id: ad016fea357f4434179632da873eec488b61af19