<R> gittemplate: update the title format
The previous is simple enough to find out which component have modify.
But there is a reason why Angular use the type prefix convention: it
helps reading one-line git commit log.

I still want to keep the simplicity, so I design the new pattern base on
their prefix:

<N> means new features
<F> means a new fix
<R> means refactor
<B> means breaking changes

This pattern marks the type of the commit. And user can easily know if
the commit contains important changes. For developer, they can easily
generate changelogs based on this pattern.

Here is example about what the new design looks like:
<N> io: create new API for better file io
<F> net/route: fix the /api/v1/users return result
<R> ci: reduce the complexity of integration test
<B> config: rename the "local_ip" field to "listen_at"

The disadvantages about this pattern design is that you need to put the
descriptions about the meaning at README file. But peoples can gain
readability. So this is a good trade-off.

Signed-off-by: Avimitin <avimitin@gmail.com>

Former-commit-id: 55084a76e814efff38cdab1d00b50f814021e0be [formerly 418b41711a44461fc664a2566d72c99a6316327b]
Former-commit-id: b1564b5ecb0a33c9741e58b0e40f6c5e3c8e2453